// src/app/modules/staff/staff-admissions/model/job-application-view-model.ts
 /**
 * Teacher API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ApplicantEducationDetailsView } from './applicant-education-details-view';
import { ApplicantExperienceDetailsView } from './applicant-experience-details-view';
import { ApplicantTrainingDetailsView } from './applicant-training-details-view';
import { AttachmentViewModel } from './attachment-view-model';
import { JobApplicationPersonView } from './job-application-person-view';



/**
 * 
 */
export interface JobApplicationViewModel { 
    /**
     * Gets or sets
     */
    seekingPosition?: string;
    /**
     * Gets or sets
     */
    passportFullName?: string;
    /**
     * Gets or sets
     */
    passportNumber?: string;
    /**
     * Gets or sets
     */
    passportDateOfIssue?: Date;
    /**
     * Gets or sets
     */
    passportPlaceOfIssue?: string;
    /**
     * Gets or sets
     */
    passportExpiryDate?: Date;
    /**
     * Gets or sets
     */
    passportIssueCountryId?: number;
    /**
     * Gets or sets
     */
    workExperienceYear?: number;
    /**
     * Gets or sets
     */
    workExperienceMonth?: number;
    /**
     * Gets or sets
     */
    specialityQualification?: string;
    /**
     * Gets or sets
     */
    reference?: string;
    /**
     * gets or sets Attachment
     */
    attachment?: AttachmentViewModel;
    /**
     * Gets or sets
     */
    applicationPerson?: JobApplicationPersonView;
    /**
     * Gets or sets
     */
    applicantEducationDetails?: Array<ApplicantEducationDetailsView>;
    /**
     * Gets or sets
     */
    applicantExperienceDetails?: Array<ApplicantExperienceDetailsView>;
    /**
     * Gets or sets
     */
    applicantTrainingDetails?: Array<ApplicantTrainingDetailsView>;
}